First correction after CodeGuru review suggestions
